WIP: Move functions to `util.ts`
